Use "raw" blockstate copy. Fixes issues with blocks that change neighbors.
An example of this are rails, the order in which they got added would
affect which way they end up facing/turning if the onBlockAdded is called
when setting the block state. Thus we don't want to call onBlockAdded when
we are trying to get an identical copy of the block states from the real
world to the fake world.